Louma/requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17.2
aniso8601==7.0.0
asgiref==3.3.4
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
branca==0.4.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
charset-normalizer==2.0.4
cloudpickle==2.0.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.11.0
Cython==0.29.26
daphne==3.0.2
debugpy==1.5.1
decorator==5.1.1
dj-database-url==0.5.0
Django==3.2.3
django-bootstrap4==21.2
django-channels==0.7.0
django-channels-graphql-ws==0.8.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.4
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-shopping-cart==0.1
djangorestframework==3.12.4
djongo==1.3.6
docopt==0.6.2
executing==0.8.2
folium==0.12.1.post1
fonttools==4.28.3
future==0.18.2
generic==1.0.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipykernel==6.9.1
ipython==8.1.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
Js2Py==0.71
jupyter-client==6.1.12
jupyter-core==4.9.2
keras==2.6.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
msgpack==1.0.2
multidict==5.1.0
mysql-connector-python==8.0.26
mysqlclient @ file:///C:/Users/Diaw%20Fara/pipwin/mysqlclient-1.4.6-cp38-cp38-win32.whl
nest-asyncio==1.5.4
numpy==1.21.4
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==21.0
parso==0.8.3
pefile==2021.9.3
pickleshare==0.7.5
Pillow==8.4.0
pipwin==0.5.1
postgis==1.0.4
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.18.0
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/Diaw%20Fara/pipwin/PyAudio-0.2.11-cp38-cp38-win32.whl
pycodestyle==2.8.0
pycparser==2.20
pygame==2.0.0.dev20
Pygments==2.11.2
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyjsparser==2.7.1
PyJWT==1.7.1
pymongo==3.12.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2021.1
pywin32==302
pywin32-ctypes==0.2.0
pyzmq==22.3.0
requests==2.25.1
requests-oauthlib==1.3.0
Rx==1.6.1
service-identity==21.1.0
setuptools-scm==6.3.2
simplejson==3.17.6
singledispatch==3.6.2
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
spyder-kernels==2.1.3
sqlparse==0.4.1
stack-data==0.2.0
stripe==2.63.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
tzdata==2021.1
tzlocal==3.0
uritemplate==4.1.1
urllib3==1.26.5
wcwidth==0.2.5
whitenoise==6.1.0
yarl==1.6.3
zope.interface==5.4.0
